#6096 setup.py sdist installation - do not create symlink schemas if exists
git-svn-id: file:///data/svn/enum/client/trunk@19077 b60c293d-5e0d-0410-87b2-fb9704ccd4b5

+++ setup.py
@@ -53,13 +53,19 @@ class EPPClientSDist(sdist):
 
     def run(self):
         "run main process"
+
         # create temporary symlink
         link_name = os.path.join(self.srcdir, 'fred', 'schemas')
-        os.symlink(EPP_SCHEMAS_PATH, link_name)
+        if os.path.exists(link_name):
+            exists = True
+        else:
+            exists = False
+            os.symlink(EPP_SCHEMAS_PATH, link_name)
         # make sdist job
         sdist.run(self)
-        # remove temporary symlink
-        os.unlink(link_name)
+        if not exists:
+            # remove temporary symlink
+            os.unlink(link_name)
 
 
 class EPPClientInstall(install):
